Added a prototype for performing a topologically-aware miRNA-mRNA integrative analysis (TAMMIA). A new class for storing the results of this analysis has been added. In particular, different helper functions has been created to: - add miRNA-mRNA interactions to biological pathways; - normalize augmented pathways; - calculate pathway score and its statistical significance through a permutation test.
Notably, to boost performance, the function for computing pathway scores has been implemented in C++ through Rcpp; and, additionally, parallel computing was included via BiocParallel.

WARNING: documentation of all functions is missing, as well as the implementation of getters/setters. Visualization functions will also have to be rewritten. Finally, there are some bugs in this prototype that still need attention.